Fix corrupted Hammy Conditions
