+++ Assets/SisGame/Characters/Hammy/Data/Condition/Hammy Not LOS of Player or Within 20f.asset
@@ -8,8 +8,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: e49cb048ac34aca4fb225c9529c115e3, type: 3}
   m_Name: Hammy Not LOS of Player or Within 20f
-  m_EditorClassIdentifier: Assembly-CSharp:SIS.Characters.Ham:HammyNegatedCondition
+  m_EditorClassIdentifier: 
   description: 
   NegatedCondition: {fileID: 11400000, guid: ef8d861ca287e30409c5166a89c9c145, type: 2}
