+++ Assets/SisGame/Characters/Hammy/Data/State Actions/Behaviour Tree/Hammy Set Exit State False.asset
@@ -8,8 +8,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 5e08eec5b68e6774aaf072dd4c568bbe, type: 3}
   m_Name: Hammy Set Exit State False
-  m_EditorClassIdentifier: Assembly-CSharp:SIS.Characters.Ham:HammySetCondtion
-  SetTo: 0
-  setableCondition: {fileID: 11400000, guid: c777e6795315d434e9ec5ef52c09cea7, type: 2}
+  m_EditorClassIdentifier: 
+  SetValue: 0
+  condition: {fileID: 11400000, guid: c777e6795315d434e9ec5ef52c09cea7, type: 2}
