+++ Assets/SisGame/Characters/Hammy/Scripts/Conditions/HammyAndCondition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1757d48550979049a788d32ed1a7c2d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
