+++ Assets/SisGame/Characters/Hammy/Scripts/Conditions/HammyNegatedCondition.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SIS.Characters.Ham
+{
+	[CreateAssetMenu(menuName = "Characters/Hammy/Conditions/HammyNegatedCondition")]
+	public class HammyNegatedCondition : HammyCondition
+	{
+		public HammyCondition NegatedCondition;
+        public override bool CheckCondition(Hammy owner)
+		{
+			return !NegatedCondition.CheckCondition(owner);
+		}
+	}
+}
