+++ Assets/SisGame/Characters/Hammy/Scripts/Conditions/HammyNegatedCondition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e49cb048ac34aca4fb225c9529c115e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
