+++ Assets/SisGame/Characters/Hammy/Scripts/State Actions/BehaviorTree/Action/HammyMoveTowardsPlayer.cs
@@ -24,4 +24,3 @@ namespace SIS.Characters.Ham
         }
     }
 }
-
