+++ Assets/SisGame/Characters/Hammy/Scripts/State Actions/BehaviorTree/Action/HammySetCondition.cs
@@ -1,19 +1,19 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
-using UnityEngine;
 using AivoTree;
+using UnityEngine;
 
 namespace SIS.Characters.Ham
 {
-    [CreateAssetMenu(menuName = "Characters/Hammy/State Actions/Behavior Tree/HammySetCondition")]
-    public class HammySetCondtion : HammyBTAction
-    {
-		public bool SetTo;
-		public HammySetableCondition setableCondition;
-        public override AivoTreeStatus Act(float timeTick, Hammy owner)
-        {
-			setableCondition.Value = SetTo;
+	[CreateAssetMenu(menuName = "Characters/Hammy/State Actions/Behavior Tree/HammySetCondition")]
+	public class HammySetCondition : HammyBTAction
+	{
+		public bool SetValue;
+		public HammySetableCondition condition;
+		public override AivoTreeStatus Act(float timeTick, Hammy owner)
+		{
+			condition.Value = SetValue;
 			return AivoTree.AivoTreeStatus.Success;
-        }
-    }
+		}
+	}
 }
